requirements.txt
numpy
scipy
opencv-python
glfw
PyGLM
PyOpenGL
imgui
tqdm
torch
mitsuba
drjit
matplotlib
open3d
